refactor: 경로 별칭 "@/*" 제거 경로 업데이트
