+++ components/Header.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import Logo from "@/components/Logo";
+import Logo from "components/Logo";
 import "twin.macro";
 
 export default function Header(): JSX.Element {
